--- a/features/step-definitions/hooks.steps.js
+++ b/features/step-definitions/hooks.steps.js
@@ -57,6 +57,7 @@ function getVariantConfig (variantTag) {
 }
 
 Before(kitStartTimeout, async function (scenario) {
+  console.log('')
   const tagNames = scenario.pickle.tags.map(x => x.name)
   const variantTags = tagNames.filter(x => x.endsWith('-variant') || x.startsWith('@kit-update'))
   if (variantTags.length > 1) {