--- a/lib/build.js
+++ b/lib/build.js
@@ -2,7 +2,7 @@
 const path = require('path')
 
 // npm dependencies
-const fse = require('fs-extra')
+const fsp = require('fs/promises')
 const fs = require('node:fs')
 const sass = require('sass')
 
@@ -37,17 +37,20 @@ function generateNunjucksSync () {
   plugins.setPluginsByType()
 }
 
-function generateAssetsSync () {
+async function generateAssets () {
   verboseLog('************************ GENERATE ASSETS ***************************')
   const timer = startPerformanceTimer()
   plugins.setPluginsByType()
   clean()
-  sassPlugins()
-  proxyUserSassIfItExists('application.scss')
-  proxyUserSassIfItExists('settings.scss')
-
-  generateCssSync()
-  endPerformanceTimer('generateAssetsSync', timer)
+  await Promise.all([
+    sassPlugins(),
+    proxyUserSassIfItExists('application.scss'),
+    proxyUserSassIfItExists('settings.scss')
+  ])
+  endPerformanceTimer('generateAssets (without css)', timer)
+  const timer2 = startPerformanceTimer()
+  await generateCss()
+  endPerformanceTimer('generateAssets (just css)', timer2)
 }
 
 function clean () {
@@ -65,9 +68,9 @@ function cleanScss () {
   } catch (e) {}
 }
 
-function ensureTempDirExists (dir = tmpDir) {
-  fse.ensureDirSync(dir, { recursive: true })
-  fse.writeFileSync(path.join(tmpDir, '.gitignore'), '*')
+async function ensureTempDirExists (dir = tmpDir) {
+  await fsp.mkdir(dir, { recursive: true })
+  await fsp.writeFile(path.join(tmpDir, '.gitignore'), '*')
 }
 
 function sassInclude (filePath) {
@@ -93,7 +96,7 @@ function sassVariables () {
   }).join('\n')
 }
 
-function sassPlugins () {
+async function sassPlugins () {
   const timer = startPerformanceTimer()
 
   const fileContents = sassVariables() +
@@ -101,38 +104,40 @@ function sassPlugins () {
       .map(sassInclude)
       .join('\n')
 
-  ensureTempDirExists(tmpSassDir)
+  await ensureTempDirExists(tmpSassDir)
 
-  fse.writeFileSync(path.join(tmpSassDir, '_plugins.scss'), fileContents)
+  await fsp.writeFile(path.join(tmpSassDir, '_plugins.scss'), fileContents)
   endPerformanceTimer('sassPlugins', timer)
 }
 
-function proxyUserSassIfItExists (filename) {
+async function proxyUserSassIfItExists (filename) {
   const timer = startPerformanceTimer()
   const userFilePath = path.join(projectDir, 'app', 'assets', 'sass', filename)
   const proxyFilePath = path.join(tmpSassDir, 'user', filename)
   const proxyFileLines = []
-  if (fse.existsSync(userFilePath)) {
+  if (fs.existsSync(userFilePath)) {
     proxyFileLines.push(sassInclude(userFilePath))
   }
-  ensureTempDirExists(path.dirname(proxyFilePath))
+  await ensureTempDirExists(path.dirname(proxyFilePath))
 
-  fse.writeFileSync(path.join(proxyFilePath), proxyFileLines.join('\n'))
+  await fsp.writeFile(path.join(proxyFilePath), proxyFileLines.join('\n'))
   endPerformanceTimer('proxyUserSassIfItExists', timer)
 }
 
-function _generateCssSync (sassPath, cssPath, options = {}) {
+async function _generateCss (sassPath, cssPath, options = {}) {
+  const timer = startPerformanceTimer()
   const { filesToSkip = [], filesToRename = {} } = options
-  if (!fse.existsSync(sassPath)) return
-  fse.mkdirSync(cssPath, { recursive: true })
-  fse.readdirSync(sassPath)
+  if (!fs.existsSync(sassPath)) return
+  await fsp.mkdir(cssPath, { recursive: true })
+  await Promise.all((await fsp.readdir(sassPath))
     .filter(file => ((
       file.endsWith('.scss') &&
       !file.startsWith('_') &&
       !filesToSkip.includes(file)
     )))
-    .forEach(file => {
+    .map(async file => {
       try {
+        const timer = startPerformanceTimer()
         const result = sass.compile(path.join(sassPath, file), {
           quietDeps: true,
           loadPaths: [projectDir],
@@ -140,9 +145,10 @@ function _generateCssSync (sassPath, cssPath, options = {}) {
           sourceMapIncludeSources: true,
           style: 'expanded'
         })
+        endPerformanceTimer('sass.compile', timer)
 
         const cssFilename = filesToRename[file] || file.replace('.scss', '.css')
-        fse.writeFileSync(path.join(cssPath, cssFilename), result.css)
+        await fsp.writeFile(path.join(cssPath, cssFilename), result.css)
       } catch (e) {
         events.emit(KIT_SASS_ERROR, {
           fileCompiling: file,
@@ -153,26 +159,27 @@ function _generateCssSync (sassPath, cssPath, options = {}) {
           }
         })
       }
-    })
+    }))
+  endPerformanceTimer('_generateCss', timer)
 }
 
-function generateCssSync (state) {
+async function generateCss (state) {
   verboseLog('************************ GENERATE SASS ***************************')
   verboseLog(state)
   const timer = startPerformanceTimer()
-  _generateCssSync(libSassDir, publicCssDir, libSassOptions)
-  _generateCssSync(appSassDir, publicCssDir, appSassOptions)
+  await Promise.all([
+    _generateCss(libSassDir, publicCssDir, libSassOptions),
+    _generateCss(appSassDir, publicCssDir, appSassOptions)
+  ])
   events.emit(KIT_SASS_REGENERATED)
-  endPerformanceTimer('generateCssSync', timer)
+  endPerformanceTimer('generateCss', timer)
 }
 
-events.on(REGENERATE_KIT_SASS, () => {
-  generateAssetsSync()
-})
+events.on(REGENERATE_KIT_SASS, generateAssets)
 
 module.exports = {
-  generateAssetsSync,
-  generateCssSync,
+  generateAssets,
+  generateCss,
   generateNunjucksSync,
   proxyUserSassIfItExists
 }