--- a/lib/config.js
+++ b/lib/config.js
@@ -1,8 +1,5 @@
-/*
- * Convenience module to make importing users app/config.js file easier,
- * handling the case where that files does not exist.
- */
-
+const { startPerformanceTimer, endPerformanceTimer } = require('./utils/performance')
+const performanceTimer = startPerformanceTimer()
 // core dependencies
 const path = require('path')
 
@@ -58,6 +55,7 @@ function getNodeEnv () {
 }
 
 function getConfig (config, swallowError = true) {
+  const timer = startPerformanceTimer()
   config = config || { ...getConfigFromFile(swallowError) }
 
   const overrideOrDefault = (configName, envName, processor, defaultValue) => {
@@ -85,9 +83,6 @@ function getConfig (config, swallowError = true) {
   overrideOrDefault('useAuth', 'USE_AUTH', asBoolean, true)
   overrideOrDefault('useHttps', 'USE_HTTPS', asBoolean, true)
   overrideOrDefault('autoReloadPages', 'AUTO_RELOAD_PAGES', asBoolean, true)
-  overrideOrDefault('logPerformance', 'LOG_PERFORMANCE', asBoolean, false)
-  overrideOrDefault('logPerformanceMatching', 'LOG_PERFORMANCE_MATCHING', asString, undefined)
-  overrideOrDefault('logPerformanceSummary', 'LOG_PERFORMANCE_SUMMARY', asNumber, undefined)
   overrideOrDefault('verbose', 'VERBOSE', asBoolean, false)
   overrideOrDefault('passwordKeys', 'PASSWORD_KEYS', asString, '')
 
@@ -116,9 +111,12 @@ function getConfig (config, swallowError = true) {
 
   config.getPluginSpecificConfig = (pluginName) => (config['plugin-specific'] || {})[pluginName] || {}
 
+  endPerformanceTimer('config lookup', timer)
+
   return config
 }
 
 module.exports = {
   getConfig
 }
+endPerformanceTimer('config setup', performanceTimer)