--- a/lib/config.test.js
+++ b/lib/config.test.js
@@ -28,7 +28,6 @@ describe('config', () => {
     passwordMissing: false,
     passwords: [],
     onGlitch: false,
-    logPerformance: false,
     nowPrototypeItAPIBaseUrl: 'https://api.nowprototype.it',
     showJsonErrors: false,
     showPrereleases: false,