We wanted to improve the performance on our hosted environment while also sharing the benefits of that improved performance.  If you're building a hosting solution for nowprototypeit we recommend that you run `npx nowprototypeit build` and then `npx nowprototypeit serve-pre-built`, this allows for the streamlined startup of the prototype in environments where the assets can't change.

This project started life as a fork, it feels appropriate to compare the performance of the production mode for the project we forked from (govuk-prototype-kit), the current live version of nowprototypeit, and the version in this pull request:

 - `govuk-prototype-kit` running `serve`: 1984.175ms average
 - `nowprototypeit@0.11.3` running `serve`: 730.2ms average
 - This pull request version running `serve`: 671.75ms average

The difference becomes even bigger when using pre-built assets (as we do on our hosting environment):

 - `nowprototypeit@0.11.3` running `serve-pre-built`: 293.175ms average
 - This pull request version running `serve-pre-built`: 128.875ms average

These figures mean there's a 93.5% performance improvement running `npx nowprototypeit serve-pre-built` when compared with the closest equivalent from before we created this fork `npx govuk-prototype-kit serve`.  Performance isn't everything, we've added a whole load of new features and wider compatibility alongside improving both local and hosted performance.

These comparisons are made on a linux laptop which is modern but not over-powered, in each case 40 runs were performed and the averages were taken. The nowprototypeit tests use the govuk-frontend-adaptor in order to have a more realistic starting point.

In order to retain these results over time a performance test has been added to the GitHub workflow which measures the performance against `nowprototypeit@0.11.3` and expects a 50% improvement for `serve-pre-built`, 4% improvement for `serve` and 30% improvement for `dev`.  These are intended to be used to avoid accidental performance decreases, if a new feature requires a slower start we may choose to accept decreasing these targets, any changes to the targets will be in the codebase and *should* also be highlighted in the PR that changes them.



The output from a local run (on a moderately powerful linux laptop) is:

```
 - - - - - - - - - - - - - - - - - 

Successes

preBuilt: Succeeded with a 61.52% improvement
serve: Succeeded with a 8.24% improvement
dev: Succeeded with a 49.23% improvement

 - - - - - - - - - - - - - - - - - 

The most important metric in understanding the goal here - this version of pre-built-serve has a [84.02]% performance improvement over the baseline version (0.11.2) to run serve.

The pre-built-serve is designed for our dedicated hosting environment but we love open-source so it's available to anyone who wants to build their own custom hosted environment.
```

The output from a very under-powered GitHub workflow runner is:

```
- - - - - - - - - - - - - - - - - 

Successes

preBuilt: Succeeded with a 63.25% improvement
serve: Succeeded with a 7.62% improvement
dev: Succeeded with a 37.38% improvement

 - - - - - - - - - - - - - - - - - 

The most important metric in understanding the goal here - this version of pre-built-serve has a [87.55]% performance improvement over the baseline version (0.11.2) to run serve.

The pre-built-serve is designed for our dedicated hosting environment but we love open-source so it's available to anyone who wants to build their own custom hosted environment.
```

To try it for yourself, checkout the `nowprototypeit` project and run:

`npm run test:perf`

Feel free to post your output as a comment on this PR.